requirements.txt
google-cloud-aiplatform>=1.13.0
google-cloud-storage>=1.44.0
mlflow>=1.30.0
scikit-learn>=1.0.2
numpy>=1.21.5
pandas>=1.3.5
fastapi>=0.95.0
uvicorn>=0.21.1
gunicorn>=20.1.0
Jinja2>=3.0.0
PyYAML>=5.1 